refactor(alias): consolidate alias commands into a single ft_alias.cc
Merge ft_alias_add.cc, ft_alias_del.cc, ft_alias_update.cc, and ft_alias_list.cc into one file. Extract shared logic into static helpers (RejectIfMultiExecInCme, FanoutAliasExists, FanoutAliasRemoved, ReplicateIfNeeded) to eliminate duplication.

Inline the consistency fanout classes from ft_alias_consistency.h
directly into ft_alias.cc and delete the header. Remove the coordinator-level unit tests (AliasConsistencyRequestTest, InfoResponseConsistencyTest) from the alias property test since they test cluster plumbing rather than alias behavior.
